判断项目名为空

+++ app/src/main/java/edu/upc/mishu/ui/activities/ModifyPssswordActivity.java
@@ -6,7 +6,10 @@ import android.util.Log;
 import android.widget.Button;
 import android.widget.EditText;
 
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.xuexiang.xutil.common.StringUtils;
 
 import java.util.List;
 
@@ -73,23 +76,27 @@ public class ModifyPssswordActivity extends AppCompatActivity {
             passwordRecord.setPassword(password.getText().toString());
             passwordRecord.setNote(note.getText().toString());
 
-            passwordRecord.encode(App.encoder,1);
+            if(StringUtils.isEmpty(passwordRecord.getName().toString())||StringUtils.isEmptyTrim(passwordRecord.getName().toString())||(passwordRecord.getName().toString().indexOf(" "))!=-1){
+                //Toast.makeText(getApplicationContext(), "项目名不能为空", Toast.LENGTH_SHORT).show();
+                AlertDialog alertDialog1 = new AlertDialog.Builder(this)
+                        .setTitle("提示")//标题
+                        .setMessage("项目名不能为空或者包含空格")//内容
+                        .create();
+                alertDialog1.show();
+            }else {
+                passwordRecord.encode(App.encoder,1);
+                passwordRecordOne.delete();
+                passwordRecord.save();
 
-            passwordRecordOne.delete();
-            passwordRecord.save();
-
-            Log.e(TAG, "onClick: pr"+passwordRecord.toString() );
-            Intent intent = new Intent();
-            intent.putExtra("name",passwordRecord.getName());
-            intent.putExtra("username",passwordRecord.getUsername());
-            intent.putExtra("id",id);
-            setResult(RESULT_OK,intent);
-            finish();
-            passwordRecordList = PasswordRecord.listAll(PasswordRecord.class);
-            for(PasswordRecord p1:passwordRecordList){
-                p1.decode(App.encoder,1);
-                Log.i("修改后密码值",p1.toString());
+                Log.e(TAG, "onClick: pr"+passwordRecord.toString() );
+                Intent intent = new Intent();
+                intent.putExtra("name",passwordRecord.getName());
+                intent.putExtra("username",passwordRecord.getUsername());
+                intent.putExtra("id",id);
+                setResult(RESULT_OK,intent);
+                finish();
             }
+
         });
     }
 }
